+++ lib/Crypto/Cipher/_mode_ccm.py
@@ -35,8 +35,9 @@ Counter with CBC-MAC (CCM) mode.
 __all__ = ['CcmMode']
 
 import struct
+from binascii import unhexlify
 
-from Crypto.Util.py3compat import (byte_string, bord, unhexlify,
+from Crypto.Util.py3compat import (byte_string, bord,
                                    _copy_bytes, _is_mutable)
 
 from Crypto.Util.strxor import strxor