+++ lib/Crypto/Cipher/_mode_eax.py
@@ -35,8 +35,9 @@ EAX mode.
 __all__ = ['EaxMode']
 
 import struct
+from binascii import unhexlify
 
-from Crypto.Util.py3compat import byte_string, bord, unhexlify, _copy_bytes
+from Crypto.Util.py3compat import byte_string, bord, _copy_bytes
 
 from Crypto.Util.strxor import strxor
 from Crypto.Util.number import long_to_bytes, bytes_to_long