+++ lib/Crypto/Cipher/_mode_gcm.py
@@ -34,7 +34,9 @@ Galois/Counter Mode (GCM).
 
 __all__ = ['GcmMode']
 
-from Crypto.Util.py3compat import byte_string, bord, unhexlify, _copy_bytes
+from binascii import unhexlify
+
+from Crypto.Util.py3compat import byte_string, bord, _copy_bytes
 
 from Crypto.Util.number import long_to_bytes, bytes_to_long
 from Crypto.Hash import BLAKE2s