+++ lib/Crypto/SelfTest/Cipher/test_CCM.py
@@ -29,9 +29,10 @@
 # ===================================================================
 
 import unittest
+from binascii import unhexlify
 
 from Crypto.SelfTest.st_common import list_test_cases
-from Crypto.Util.py3compat import unhexlify, tobytes, bchr, b, _memoryview
+from Crypto.Util.py3compat import tobytes, bchr, _memoryview
 from Crypto.Cipher import AES
 from Crypto.Hash import SHAKE128
 
@@ -76,7 +77,7 @@ class CcmTests(unittest.TestCase):
 
     def test_nonce_length(self):
         self.assertRaises(ValueError, AES.new, self.key_128, AES.MODE_CCM,
-                          nonce=b(""))
+                          nonce=b"")
         self.assertRaises(ValueError, AES.new, self.key_128, AES.MODE_CCM,
                           nonce=bchr(1) * 6)
         self.assertRaises(ValueError, AES.new, self.key_128, AES.MODE_CCM,
@@ -112,17 +113,17 @@ class CcmTests(unittest.TestCase):
     def test_null_encryption_decryption(self):
         for func in "encrypt", "decrypt":
             cipher = AES.new(self.key_128, AES.MODE_CCM, nonce=self.nonce_96)
-            result = getattr(cipher, func)(b(""))
-            self.assertEqual(result, b(""))
+            result = getattr(cipher, func)(b"")
+            self.assertEqual(result, b"")
 
     def test_either_encrypt_or_decrypt(self):
         cipher = AES.new(self.key_128, AES.MODE_CCM, nonce=self.nonce_96)
-        cipher.encrypt(b(""))
-        self.assertRaises(TypeError, cipher.decrypt, b(""))
+        cipher.encrypt(b"")
+        self.assertRaises(TypeError, cipher.decrypt, b"")
 
         cipher = AES.new(self.key_128, AES.MODE_CCM, nonce=self.nonce_96)
-        cipher.decrypt(b(""))
-        self.assertRaises(TypeError, cipher.encrypt, b(""))
+        cipher.decrypt(b"")
+        self.assertRaises(TypeError, cipher.encrypt, b"")
 
     def test_data_must_be_bytes(self):
         cipher = AES.new(self.key_128, AES.MODE_CCM, nonce=self.nonce_96)
@@ -172,7 +173,7 @@ class CcmTests(unittest.TestCase):
         # More than zero
         cipher = AES.new(self.key_128, AES.MODE_CCM, nonce=self.nonce_96,
                          assoc_len=0)
-        self.assertRaises(ValueError, cipher.update, b("1"))
+        self.assertRaises(ValueError, cipher.update, b"1")
 
         # Too large
         cipher = AES.new(self.key_128, AES.MODE_CCM, nonce=self.nonce_96,
@@ -254,7 +255,7 @@ class CcmTests(unittest.TestCase):
 
             for chunk in break_up(auth_data, chunk_length):
                 cipher.update(chunk)
-            pt2 = b("")
+            pt2 = b""
             for chunk in break_up(ciphertext, chunk_length):
                 pt2 += cipher.decrypt(chunk)
             self.assertEqual(plaintext, pt2)
@@ -268,7 +269,7 @@ class CcmTests(unittest.TestCase):
 
             for chunk in break_up(auth_data, chunk_length):
                 cipher.update(chunk)
-            ct2 = b("")
+            ct2 = b""
             for chunk in break_up(plaintext, chunk_length):
                 ct2 += cipher.encrypt(chunk)
             self.assertEqual(ciphertext, ct2)
@@ -457,8 +458,8 @@ class CcmFSMTests(unittest.TestCase):
     def test_valid_multiple_encrypt_or_decrypt(self):
         # Only possible if msg_len is declared in advance
         for method_name in "encrypt", "decrypt":
-            for auth_data in (None, b("333"), self.data_128,
-                              self.data_128 + b("3")):
+            for auth_data in (None, b"333", self.data_128,
+                              self.data_128 + b"3"):
                 if auth_data is None:
                     assoc_len = None
                 else: